+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `DisputeAccepted`, `DisputeDisputed`, and `DisputeExpired` webhook constants
 * Add `Authentication Insight` support to payment method nonce create
 * Add `recurringCustomerConsent` and `recurringMaxAmount` to `authenticationInsightOptions` for `PaymentMethodNonce.create`
+* Add `FileIsEmpty` error code
 
 ## 2.22.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors